requirements.txt
astroid==2.4.2
autopep8==1.5.4
click==7.1.2
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
numpy==1.19.4
opencv-contrib-python==4.4.0.46
Pillow==8.0.1
pycodestyle==2.6.0
pylint==2.6.0
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyserial==3.4
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
wrapt==1.12.1
